add delete news from main
